requirements.txt
numpy
scipy
matplotlib
Pillow
pyueye
automation1
